added missing tooltip text move the UI component in the UI package
