+++ src/main/java/com/swissas/dialog/DtoGeneratorForm.jfd
@@ -13,7 +13,7 @@ new FormModel {
 			add( new FormContainer( "javax.swing.JPanel", new FormLayoutManager( class net.miginfocom.swing.MigLayout ) {
 				"$layoutConstraints": "insets 0,hidemode 3,gap 10 5"
 				"$columnConstraints": "[grow,fill]"
-				"$rowConstraints": "[fill][fill][fill][fill][fill][fill][fill][grow,fill]"
+				"$rowConstraints": "[][fill][fill][fill][fill][fill][fill][grow,fill]"
 			} ) {
 				name: "panel1"
 				auxiliary() {
